remove mktuntap for crosvm
The current crosvm runner uses mktuntap to pass fds of the
network devices to a vm. It seems to silently ignore declared
mac addresses as well as non tap-devices.

Current versions of crosvm in nixpkgs also support a
--net flag where tap id & mac can be passed directly, so I
believe mktuntap is not needed (anymore)? Please let me
know if it still provides value.