cloud-hypervisor: resolve issue #152 using waitpid
The systemd unit ExecStop script for cloud-hypervisor used to wait for
machine termination by detecting the closing of the cloud-hypervisor
API socket. However, it seems that this socket is actually closed
before the machine is safely shut down.

Therefore, this patch replaces the listen on the socket with a waitpid
on the main PID, which is the cloud-hypervisor process. This process
only terminates after the machine has properly shut down.