feat: allow to specify a CPU to emulate with a cross-system
By setting a module of the form:

```
modules = [
  { nixpkgs.crossSystem.config = "aarch64-unknown-linux-gnu"; }
]
```

On the `lib.nixosSystem` call, and a specific CPU on the
`nixosModules.microvm` call, such as:

```
microvm = {
  cpu = "cortex-a53";
}
```

Allow a different CPU to be emulated by using qemu's Tiny Code
Generator (TCG). While this will have a performance impact on the
guest, it allows to experiment cross-system derivations on an emulated
guest.